Cleanup an unintended sign extension warning in voodoo (CWE-194)
Fixes the following warning flagged by Coverity:

  The sign-extended value may be interpreted as a very large
  number. In raster_fastfill(void *, int, poly_extent const *,
  unsigned short const *): Value may be sign extended
  unintentionally (CWE-194)

+++ src/hardware/voodoo.cpp
@@ -3623,7 +3623,8 @@ static void raster_fastfill(void *destbase, int32_t y, const poly_extent *extent
 
 		const uint64_t expanded = (static_cast<uint64_t>(color) << 48) |
 		                          (static_cast<uint64_t>(color) << 32) |
-		                          (color << 16) | color;
+		                          (static_cast<uint32_t>(color) << 16) |
+		                          color;
 
 		uint16_t* dest = reinterpret_cast<uint16_t*>(v->fbi.ram +
 		                                             v->fbi.auxoffs) +
